backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
transformers
torch
scikit-learn
spacy
pypdf # <-- TAMBAHKAN INI
